requirements.txt
# Data processing and validation
pydantic>=2.0.0
numpy>=1.21.0
sqlalchemy>=2.0.0

# LLM and AI integration
openai>=1.0.0
anthropic>=0.21.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
kor>=0.3.0

# PDF processing
pypdf>=3.0.0
pdfplumber>=0.9.0
pymupdf>=1.23.0
pdfminer.six>=20221105
pytesseract>=0.3.0
pdf2image>=1.16.0

# Document processing
python-docx>=0.8.11
mammoth>=0.3.0

# Azure services
azure-core>=1.26.0
azure-ai-documentintelligence>=1.0.0

# HTTP and requests
requests>=2.26.0

# Optional: Environment and configuration (from lineage requirements)
python-dotenv>=0.19.0
PyYAML>=5.4.0
python-dateutil>=2.8.0


